// studio/src/api/generated/ChatMessageIn.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessageContent } from "./MessageContent";
import type { ToolCall } from "./ToolCall";

export type ChatMessageIn = { role: string, content?: MessageContent | null, tool_calls?: Array<ToolCall> | null, tool_call_id?: string | null, name?: string | null, };